config/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dj-database-url==3.1.2
Django==6.0.1
django-anymail==14.0
django-celery-beat==2.9.0
django-filter==25.2
django-restframework==0.0.1
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.14
fastapi==0.128.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-mock==3.15.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
starlette==0.50.0
stripe==14.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
